fix linespace bug
https://github.com/ITTcs/fuzzylab/issues/2

+++ fuzzylab/linspace.py
@@ -5,4 +5,4 @@
 import numpy as np
 
 def linspace(start, step, stop):
-    return np.linspace(start, stop, (abs(stop - start) / step) + 1)
+    return np.linspace(start, stop, int(abs(stop - start) / step) + 1)
